clipboard-apis/clipboard_events_relative_to_other_input_events_cut_on_inputtext.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - clipboard events relative to other input events   - cut on 'input_text'</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** clipboard events relative to other input events   */
/* Events: cut */
/* Test HTML: <form action=""><input id="input_text" autofocus onfocus="this.select()" value="copied text" required></form>*/
/* Targets: 'input_text' */

var eventsFired=[];
function logEvent(e){eventsFired.push(e.type);};
document.addEventListener('DOMContentLoaded', function(){
  document.getElementById('input_text').addEventListener( 'change', logEvent, false );
  document.getElementById('input_text').addEventListener( 'input', logEvent, false );
  document.getElementById('input_text').addEventListener( 'invalid', logEvent, false );
  document.getElementById('input_text').addEventListener( 'textInput', logEvent, false );
},false);

function clipboard_api_test(e, test_obj){
  logEvent(e);
  test_obj.step_timeout(function(){
    e.target.blur(); // causes change event
    e.target.form.checkValidity(); // causes invalid event
  }, 10);
  test_obj.step_timeout(function(){
    var expected = ['cut','cut handler done','input','change','invalid']; // Issue: should textInput fire? Chrome doesn't but I don't know if it's on purpose
    result(function(){
      assert_array_equals(eventsFired, expected);
    });
  }, 500);
  logEvent( {type: e.type+' handler done'} );
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <form action=""><input id="input_text" autofocus onfocus="this.select()" value="copied text" required></form>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste='';
        var extPassCond=null;
        var eventTarget='input_text';
        var eventType='cut';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
